Change state on the EmailInput
For a better UX, if the user touch the button without type his email and
password, the inputs' borders are going to change their color to red.
So, after that all the fields will be red and the user will type them on
that color. With this change if the user types a validate email but a
wrong password, the email's field will return at his original color and
the password's field will keep in red color.
Also I delete unnecessary comments